pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.127.0",
    "fastapi-users[sqlalchemy]>=14.0.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "supabase>=2.27.0",
    "uvicorn[standard]>=0.39.0",
]
